feat: creados endpoints de login y register completos PENDIENTE APLICAR SEGURIDAD

+++ src/main/java/com/ferrinsa/fairpartner/user/dto/UserAdminDTO.java
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.NotBlank;
 import java.time.LocalDate;
 
 public record UserAdminDTO(
-        @NotBlank int id,
+        @NotBlank Long id,
         @NotBlank String name,
         @NotBlank @Email String email,
         @NotBlank String password,
